🧹 test: consolidate security_branches_test.go into security_test.go
Move every test, helper, and type from security_branches_test.go into
security_test.go and delete the branches file. No test logic changes —
purely a file merge:

- Union the imports (add errors, net/url, time to security_test.go).
- Append the branch-coverage / edge-case tests and their shared
  redirect-client helpers (redirectStep, countingRedirectClient,
  newCountingRedirectClient, roundTripperFunc, mustParseTestURL) under
  a section separator.
- No name collisions; both files were already package proxy and
  referenced each other's helpers.

Proxy test-function count is unchanged (121 across the package). Full
suite, vet, and lint stay green.
